Zv-303 - Financial Information Tutorial/Subtutorials

+++ db/migrate/20170518003442_add_description_to_tutorials.rb
@@ -1,6 +1,13 @@
 class AddDescriptionToTutorials < ActiveRecord::Migration
   def change
-    add_column :tutorials, :description, :string
-    remove_column :tutorials, :slug
+    unless column_exists? :tutorials, :description
+      add_column :tutorials, :description, :string
+      Tutorial.reset_column_information
+    end
+    
+    if column_exists? :tutorials, :slug
+      remove_column :tutorials, :slug
+      Tutorial.reset_column_information
+    end
   end
 end
