+++ db/migrate/20170522212157_cleanup_stripe_subscriptions.rb
@@ -1,10 +1,20 @@
 class CleanupStripeSubscriptions < ActiveRecord::Migration
   def change
-    remove_column :stripe_subscriptions, :last4
-    remove_column :stripe_subscriptions, :name_on_card
-    remove_column :stripe_subscriptions, :stripe_token
+    if column_exists? :stripe_subscriptions, :last4
+      remove_column :stripe_subscriptions, :last4
+    end
+    
+    if column_exists? :stripe_subscriptions, :name_on_card
+      remove_column :stripe_subscriptions, :name_on_card
+    end
+    
+    if column_exists? :stripe_subscriptions, :stripe_token
+      remove_column :stripe_subscriptions, :stripe_token
+    end
 
-    add_index :stripe_subscriptions, :user_id
+    unless index_exists? :stripe_subscriptions, :user_id
+      add_index :stripe_subscriptions, :user_id
+    end
 
     StripeSubscription.includes(:user).all.each do |sub|
       if sub.user.present? && sub.user.stripe_customer_record.blank?
